:recycle: Refactor: Service 계층 리팩토링 #42
