--- src/main/java/site/moamoa/backend/service/category/query/CategoryQueryService.java
@@ -1,7 +1,0 @@
-package site.moamoa.backend.service.category.query;
-
-import site.moamoa.backend.domain.Category;
-
-public interface CategoryQueryService {
-    Category findCategoryById(Long id);
-}
