+++ src/main/java/site/moamoa/backend/service/component/command/member_post/MemberPostCommandService.java
@@ -1,0 +1,8 @@
+package site.moamoa.backend.service.component.command.member_post;
+
+import site.moamoa.backend.domain.mapping.MemberPost;
+
+public interface MemberPostCommandService {
+  void save(MemberPost memberPost);
+
+}
