+++ src/main/java/site/moamoa/backend/service/component/command/member_post/MemberPostCommandServiceImpl.java
@@ -1,0 +1,17 @@
+package site.moamoa.backend.service.component.command.member_post;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+import site.moamoa.backend.domain.mapping.MemberPost;
+import site.moamoa.backend.repository.mapping.MemberPostRepository;
+
+@AllArgsConstructor
+@Service
+public class MemberPostCommandServiceImpl implements MemberPostCommandService {
+
+  private final MemberPostRepository memberPostRepository;
+  @Override
+  public void save(MemberPost memberPost) {
+    memberPostRepository.save(memberPost);
+  }
+}
