+++ src/main/java/site/moamoa/backend/service/component/command/postimage/PostImageCommandService.java
@@ -1,0 +1,5 @@
+package site.moamoa.backend.service.component.command.postimage;
+
+public interface PostImageCommandService {
+  void deletePostImageByPostId(Long id);
+}
