+++ src/main/java/site/moamoa/backend/service/component/query/category/CategoryQueryService.java
@@ -1,0 +1,7 @@
+package site.moamoa.backend.service.component.query.category;
+
+import site.moamoa.backend.domain.Category;
+
+public interface CategoryQueryService {
+    Category findCategoryById(Long id);
+}
