+++ src/main/java/site/moamoa/backend/service/component/query/member_post/MemberPostQueryService.java
@@ -1,0 +1,6 @@
+package site.moamoa.backend.service.component.query.member_post;
+
+public interface MemberPostQueryService {
+
+    void checkAuthor(Long memberId, Long postId);
+}
