--- src/main/java/site/moamoa/backend/service/memberpost/command/MemberPostCommandService.java
@@ -1,8 +1,0 @@
-package site.moamoa.backend.service.memberpost.command;
-
-import site.moamoa.backend.domain.mapping.MemberPost;
-
-public interface MemberPostCommandService {
-  void save(MemberPost memberPost);
-
-}
