--- src/main/java/site/moamoa/backend/service/memberpost/query/MemberPostQueryService.java
@@ -1,6 +1,0 @@
-package site.moamoa.backend.service.memberpost.query;
-
-public interface MemberPostQueryService {
-
-    void checkAuthor(Long memberId, Long postId);
-}
