+++ src/main/java/site/moamoa/backend/service/module/member/MemberModuleServiceImpl.java
@@ -1,0 +1,38 @@
+package site.moamoa.backend.service.module.member;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import site.moamoa.backend.api_payload.code.status.ErrorStatus;
+import site.moamoa.backend.api_payload.exception.handler.MemberHandler;
+import site.moamoa.backend.domain.Member;
+import site.moamoa.backend.domain.Post;
+import site.moamoa.backend.domain.enums.CapacityStatus;
+import site.moamoa.backend.repository.mapping.MemberPostRepository;
+import site.moamoa.backend.repository.member.MemberRepository;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class MemberModuleServiceImpl implements MemberModuleService {
+    private final MemberRepository memberRepository;
+    private final MemberPostRepository memberPostRepository;
+
+    @Override
+    public List<Post> findRecruitingMember(Long memberId, CapacityStatus status) {
+        return memberPostRepository.findRecruitingMember(memberId, status);
+    }
+
+    @Override
+    public List<Post> findParticipatedMember(Long memberId, CapacityStatus status) {
+        return memberPostRepository.findParticipatedMember(memberId, status);
+    }
+
+    @Override
+    public Member findMemberById(Long id) {
+        return memberRepository.findById(id)
+                .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
+    }
+}
