--- src/main/java/site/moamoa/backend/service/post/query/PostQueryService.java
@@ -1,18 +1,0 @@
-package site.moamoa.backend.service.post.query;
-
-import site.moamoa.backend.web.dto.response.PostResponseDTO;
-
-public interface PostQueryService {
-
-    PostResponseDTO.GetPosts findPostsByNear(Long memberId);
-
-    PostResponseDTO.GetPosts findPostsByLatest();
-
-    PostResponseDTO.GetPosts findPostsByRanking(Long memberId);
-
-    PostResponseDTO.GetPosts findPostsByRecentKeyword(Long memberId);
-
-    PostResponseDTO.GetPosts findPostsByConditions(String keyword, String category, Integer dDay, Integer total, Integer minPrice, Integer maxPrice);
-
-    PostResponseDTO.GetPost findPostById(Long postId);
-}
