--- src/main/java/site/moamoa/backend/service/postimage/command/PostImageCommandService.java
@@ -1,5 +1,0 @@
-package site.moamoa.backend.service.postimage.command;
-
-public interface PostImageCommandService {
-  void deletePostImageByPostId(Long id);
-}
