--- src/main/java/site/moamoa/backend/service/postimage/query/PostImageQueryService.java
@@ -1,8 +1,0 @@
-package site.moamoa.backend.service.postimage.query;
-
-import java.util.List;
-import site.moamoa.backend.domain.mapping.PostImage;
-
-public interface PostImageQueryService {
-    List<PostImage> findPostImageByPostId(Long id);
-}
