+++ src/main/java/site/moamoa/backend/web/controller/MemberController.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import site.moamoa.backend.api_payload.ApiResponseDTO;
 import site.moamoa.backend.domain.enums.CapacityStatus;
-import site.moamoa.backend.service.member.command.MemberCommandService;
-import site.moamoa.backend.service.member.query.MemberQueryService;
+import site.moamoa.backend.service.component.command.member.MemberCommandService;
+import site.moamoa.backend.service.component.query.member.MemberQueryService;
 import site.moamoa.backend.web.dto.base.AuthInfoDTO;
 import site.moamoa.backend.web.dto.request.MemberRequestDTO.UpdateMemberAddress;
 import site.moamoa.backend.web.dto.response.PostResponseDTO.GetMyPostList;
